+++ src/index.ts
@@ -1,5 +1,6 @@
 export { generateNumbersArray } from './functions/generate-numbers-array';
 export { uniquifyArray, uniquifyObjectsArray } from './functions/uniquify-array';
+export { getRandomItemFromArray } from './functions/get-random-item-from-array';
 export { objectsArrayToObject } from './functions/objects-array-to-object';
 export { sortArrayByObjectProperty } from './functions/sort-array-by-object-property';
 export { groupArrayByObjectProperty } from './functions/group-array-by-object-property';